+++ school-cal/src/components/AdminDashboard/AdminAddEvent.js
@@ -1,0 +1,22 @@
+import React from 'react';
+// import FullCalendar from '@fullcalendar/react'
+// import dayGridPlugin from '@fullcalendar/daygrid'
+
+const AdminAddEvent = () => {
+    return (
+        <div className = "AdminAddEvent-container">
+            <p>select date you want to add an event on</p>
+            {/* <FullCalendar
+                defaultView="dayGridMonth"
+                plugins={[ dayGridPlugin ]}
+                weekends={false}
+                events={[
+                    { title: 'event 1', date: '2019-04-01' },
+                    { title: 'event 2', date: '2019-04-02' }
+                ]}
+            /> */}
+        </div>
+    );
+};
+
+export default AdminAddEvent;
